--- packages/tamagui-core/src/core/index.tsx
@@ -1,1 +1,0 @@
-export * from "./components";
