+++ packages/tamagui-core/src/index.tsx
@@ -1,5 +1,4 @@
 export * from "./components";
-export * from "./core";
 // @ts-expect-error
 export { useTheme as useTheme2 } from "./hooks";
 export { appConfig as config } from "./tamagui.config";
